# docs/dev/api/openapi.yaml
openapi: 3.1.0
info:
  title: Auth API
  description: 'Authentication API: accounts, sessions, second factors and client applications. Issues ES256 access tokens and rotating refresh tokens, and publishes the JWKS resource servers verify them with.'
  license:
    name: MIT
  version: 2.0.1
servers:
- url: http://localhost:3000
  description: Local development
paths:
...
  /.well-known/oauth-authorization-server:
    get:
      tags:
      - oauth
      operationId: metadata
      responses:
        '200':
          description: Authorization server metadata (RFC 8414)
          content:
            application/json:
              schema:
                type: object
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /users/me/audit:
    get:
      tags:
      - account
      summary: GET /users/me/audit - newest first.
      operationId: list
      parameters:
      - name: limit
        in: query
        description: Entries per page, 1-200 (default 50)
        required: false
        schema:
          type: integer
          format: int64
      - name: cursor
        in: query
        description: next_cursor of the previous page
        required: false
        schema:
          type: string
      responses:
        '200':
          description: The caller's security history, newest first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditPageResponse'
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: '`first_party_session_required`: a token delegated to a client or issued for a personal access token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid cursor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
...
components:
  schemas:
...
    AuditEntryResponse:
      type: object
      required:
      - id
      - created_at
      - action
      - metadata
      properties:
        action:
          type: string
          description: 'Stable snake_case name: `login`, `password_changed`, ...'
        created_at:
          type: integer
          format: int64
          description: Unix timestamp (seconds), like every other timestamp of this API.
        id:
          type: string
          format: uuid
        ip_address:
          type:
          - string
          - 'null'
        metadata:
          type: object
        request_id:
          type:
          - string
          - 'null'
          format: uuid
    AuditPageResponse:
      type: object
      required:
      - entries
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/AuditEntryResponse'
        next_cursor:
          type:
          - string
          - 'null'
          description: Pass back as `cursor` to read the next page; absent on the last one.
...
    ErrorBody:
      type: object
      description: Body of every error response.
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: 'Stable and machine-readable: `invalid_credentials`, `reauthentication_required`, ...'
        message:
          type: string
          description: Human-readable explanation; may change between versions.
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...